Moving realtime to web.realtime in compute
- Moving the realtime package under web. This is clearer especially now that there's other web resources too

- Add a unit test for the status resource
