+++ compute/src/main/java/com/chatalytics/compute/web/realtime/ComputeRealtimeServer.java
@@ -1,4 +1,4 @@
-package com.chatalytics.compute.realtime;
+package com.chatalytics.compute.web.realtime;
 
 import com.google.common.util.concurrent.AbstractIdleService;
